lnwallet: move code to transactions files
